dev: fix for misaligned FMC
This commit fixes the FMC orientation by rotating it 180 degrees.  The
original layout was created while looking at a mirrored view of an
example FMC expansion board - oops.

This update also adds a standard 20-pin 0.1" pitch Arm debug connector
that is wired for SWD only.  In addition there are 2 8-pin 0.1" headers
connected to FPGA IO that can be used for user IO of any sort.